refac: rename `xxx_traits.h` to `xxx_traits_forward.h`
For `xxx_traits.h` that contains only forward declaration for `XXXTraits`,
the filename is changed to `xxx_traits_forward.h` to deliever its
meaning more clearly. Plus, it's also conformed to other files.
