+++ benchmark/msm/BUILD.bazel
@@ -23,7 +23,7 @@ tachyon_cc_library(
     deps = [
         ":simple_msm_benchmark_reporter",
         "//tachyon/base/time",
-        "//tachyon/cc/math/elliptic_curves:point_traits",
+        "//tachyon/cc/math/elliptic_curves:point_traits_forward",
     ],
 )
