+++ tachyon/cc/math/elliptic_curves/BUILD.bazel
@@ -6,7 +6,7 @@ tachyon_cc_library(
     name = "point_conversions",
     hdrs = ["point_conversions.h"],
     deps = [
-        ":point_traits",
+        ":point_traits_forward",
         "//tachyon/cc/math/finite_fields:prime_field_conversions",
         "//tachyon/math/geometry:point2",
         "//tachyon/math/geometry:point3",
@@ -15,6 +15,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "point_traits",
-    hdrs = ["point_traits.h"],
+    name = "point_traits_forward",
+    hdrs = ["point_traits_forward.h"],
 )
