--- tachyon/cc/math/elliptic_curves/point_traits.h
@@ -1,11 +1,0 @@
-#ifndef TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_H_
-#define TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_H_
-
-namespace tachyon::cc::math {
-
-template <typename Point>
-struct PointTraits;
-
-}  // namespace tachyon::cc::math
-
-#endif  // TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_H_
