+++ tachyon/cc/math/elliptic_curves/point_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_FORWARD_H_
+#define TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_FORWARD_H_
+
+namespace tachyon::cc::math {
+
+template <typename Point>
+struct PointTraits;
+
+}  // namespace tachyon::cc::math
+
+#endif  // TACHYON_CC_MATH_ELLIPTIC_CURVES_POINT_TRAITS_FORWARD_H_
