--- tachyon/cc/math/finite_fields/prime_field_traits.h
@@ -1,11 +1,0 @@
-#ifndef TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_H_
-#define TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_H_
-
-namespace tachyon::cc::math {
-
-template <typename T>
-struct PrimeFieldTraits;
-
-}  // namespace tachyon::cc::math
-
-#endif  // TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_H_
