+++ tachyon/cc/math/finite_fields/prime_field_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_FORWARD_H_
+#define TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_FORWARD_H_
+
+namespace tachyon::cc::math {
+
+template <typename T>
+struct PrimeFieldTraits;
+
+}  // namespace tachyon::cc::math
+
+#endif  // TACHYON_CC_MATH_FINITE_FIELDS_PRIME_FIELD_TRAITS_FORWARD_H_
