--- tachyon/crypto/commitments/vector_commitment_scheme_traits.h
@@ -1,11 +1,0 @@
-#ifndef TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_H_
-#define TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_H_
-
-namespace tachyon::crypto {
-
-template <typename C>
-struct VectorCommitmentSchemeTraits;
-
-}  // namespace tachyon::crypto
-
-#endif  // TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_H_
