+++ tachyon/crypto/commitments/vector_commitment_scheme_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_FORWARD_H_
+#define TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_FORWARD_H_
+
+namespace tachyon::crypto {
+
+template <typename C>
+struct VectorCommitmentSchemeTraits;
+
+}  // namespace tachyon::crypto
+
+#endif  // TACHYON_CRYPTO_COMMITMENTS_VECTOR_COMMITMENT_SCHEME_TRAITS_FORWARD_H_
