--- tachyon/math/base/bit_traits.h
@@ -1,11 +1,0 @@
-#ifndef TACHYON_MATH_BASE_BIT_TRAITS_H_
-#define TACHYON_MATH_BASE_BIT_TRAITS_H_
-
-namespace tachyon::math {
-
-template <typename T, typename SFINAE = void>
-class BitTraits;
-
-}  // namespace tachyon::math
-
-#endif  // TACHYON_MATH_BASE_BIT_TRAITS_H_
