+++ tachyon/math/base/bit_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_MATH_BASE_BIT_TRAITS_FORWARD_H_
+#define TACHYON_MATH_BASE_BIT_TRAITS_FORWARD_H_
+
+namespace tachyon::math {
+
+template <typename T, typename SFINAE = void>
+class BitTraits;
+
+}  // namespace tachyon::math
+
+#endif  // TACHYON_MATH_BASE_BIT_TRAITS_FORWARD_H_
