+++ tachyon/math/elliptic_curves/short_weierstrass/sw_curve.h
@@ -9,7 +9,7 @@
 #include "tachyon/math/elliptic_curves/point_conversions.h"
 #include "tachyon/math/elliptic_curves/point_xyzz.h"
 #include "tachyon/math/elliptic_curves/projective_point.h"
-#include "tachyon/math/elliptic_curves/short_weierstrass/sw_curve_traits.h"
+#include "tachyon/math/elliptic_curves/short_weierstrass/sw_curve_traits_forward.h"
 
 namespace tachyon::math {
