--- tachyon/math/elliptic_curves/short_weierstrass/sw_curve_traits.h
@@ -1,11 +1,0 @@
-#ifndef TACHYON_MATH_ELLIPTIC_CURVES_SHORT_WEIERSTRASS_SW_CURVE_TRAITS_H_
-#define TACHYON_MATH_ELLIPTIC_CURVES_SHORT_WEIERSTRASS_SW_CURVE_TRAITS_H_
-
-namespace tachyon::math {
-
-template <typename T>
-struct SWCurveTraits;
-
-}  // namespace tachyon::math
-
-#endif  // TACHYON_MATH_ELLIPTIC_CURVES_SHORT_WEIERSTRASS_SW_CURVE_TRAITS_H_
