+++ tachyon/math/polynomials/BUILD.bazel
@@ -14,12 +14,12 @@ tachyon_cc_library(
         "support_poly_operators.h",
     ],
     deps = [
-        ":polynomial_traits",
+        ":polynomial_traits_forward",
         "//tachyon/math/base:ring",
     ],
 )
 
 tachyon_cc_library(
-    name = "polynomial_traits",
-    hdrs = ["polynomial_traits.h"],
+    name = "polynomial_traits_forward",
+    hdrs = ["polynomial_traits_forward.h"],
 )
