--- tachyon/math/polynomials/polynomial_traits.h
@@ -1,15 +1,0 @@
-#ifndef TACHYON_MATH_POLYNOMIALS_POLYNOMIAL_TRAITS_H_
-#define TACHYON_MATH_POLYNOMIALS_POLYNOMIAL_TRAITS_H_
-
-#include <type_traits>
-
-#include "tachyon/math/base/ring.h"
-
-namespace tachyon::math {
-
-template <typename T>
-class PolynomialTraits;
-
-}  // namespace tachyon::math
-
-#endif  // TACHYON_MATH_POLYNOMIALS_POLYNOMIAL_TRAITS_H_
