+++ tachyon/zk/base/commitments/univariate_polynomial_commitment_scheme_extension_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_ZK_BASE_COMMITMENTS_UNIVARIATE_POLYNOMIAL_COMMITMENT_SCHEME_EXTENSION_TRAITS_FORWARD_H_
+#define TACHYON_ZK_BASE_COMMITMENTS_UNIVARIATE_POLYNOMIAL_COMMITMENT_SCHEME_EXTENSION_TRAITS_FORWARD_H_
+
+namespace tachyon::zk {
+
+template <typename C>
+struct UnivariatePolynomialCommitmentSchemeExtensionTraits;
+
+}  // namespace tachyon::zk
+
+#endif  // TACHYON_ZK_BASE_COMMITMENTS_UNIVARIATE_POLYNOMIAL_COMMITMENT_SCHEME_EXTENSION_TRAITS_FORWARD_H_
